+++ RELEASENOTES.md
@@ -55,6 +55,9 @@
     *   Add new parameter to `MediaSession.Callback.onPlaybackResumption` to
         indicate if the call happens to gather information only or to start
         playback ([#1764](https://github.com/androidx/media/issues/1764)).
+    *   Add `MediaSessionService.triggerNotificationUpdate` to manually trigger
+        a notification update
+        ([#1833](https://github.com/androidx/media/issues/1833)).
 *   UI:
     *   Fix bug where `PlayerSurface` inside re-usable components like
         `LazyColumn` didn't work correctly